dag/set: provide reverse() API
Summary:
Now we track the set order. Add a way to reverse it.

Initially I used this in CodeReviewProvider APIs so it can speak `HashSet`
instead of `Array<CommitInfo>`, so the `HighlightedCommits` can take a
`HashSet` instead of creating a set from `Array<CommitInfo>` every time. Later
I found that change has pros and cons overhead-wise, and didn't pursue it.
But the `reverse` API seems useful.

Reviewed By: muirdm

Differential Revision: D52861867

fbshipit-source-id: 5b70038859e87c35e2a0f9f8667196d3452154e1