Support for pre-reqs in bundle create command
Summary:
The `create-bundle` command takes a `have-heads` argument which allows the user to specify which heads are already present with them so that the bundle can skip including those heads and their ancestors. However, in addition to skipping those commits we need to also specify them as pre-requisites in the bundle otherwise the bundle looks like it contains the full history of the repo where in reality is just contains a subset.

This diff ensures that the `have-heads` argument is used to populate the `prerequisites` part of the generated bundle.

Differential Revision: D52869325

fbshipit-source-id: 7306b32a832dd1796d9f21e03d39f847cc03eb53