dag: add an immutable `CommitInfo` type
Summary:
Making the vanilla `CommitInfo` reuse older objects are not enough for fast
equality check, since `Dag` and preview logic will extend the `CommitInfo`
with extra fields. It seems `SelfUpdate` immutable record is the way to
solve the problem (with dynamically calculated additional fields). This
diff adds a `SelfUpdate` type for that.

Reviewed By: evangrayk

Differential Revision: D52958649

fbshipit-source-id: adf64ad34290c2afd8319bb6d1d2b3efc9cd4bbb