cached: add an audit mode
Summary:
The cached function is expected to be "pure" and produces the same result on
the same input. However, sometimes this is not satisfied (and TypeScript /
Javascript have no way to express "pure"-ness). Add an "audit" mode to compare
cached value with recalculated value to help debug this kind of issues.

Note: the actual issue seems to be "SelfUpdate" mutability, not "cached".

Reviewed By: evangrayk

Differential Revision: D52927195

fbshipit-source-id: 9da3d38c33adbfbc62b842148eb3cd95fb6ebc01