DagCommitList: move commit highlight logic out of RenderDag
Summary:
Now I'm a bit more familiar with the ideas behind Recoil (or Jotai), this is
a better version of D52874611.

The problem is that `RenderDag` is intended to be "general purpose" - taking
"flexible" callbacks (ex. `renderGlyph`). And we also want highlight change to
only re-render the affected circles, not the whole dag. D52874611 solved the
latter problem but regressed the first.

But, "flexible callback" does not mean that it has to be the "default" callback
to be static, and memo-able. It could also be custom static functions which
renders components that use Recoil states, as the previous diff demonstrates.

Reviewed By: evangrayk

Differential Revision: D52958629

fbshipit-source-id: c89902a160ba6d90be45d7ab0353980e665844f2