RenderDag: handle avatar highlight at the leaf level
Summary:
A "highlighedCommits" change re-renders the entire dag tree unnecessarily.
This diff changes it to only re-render leaf nodes (`RegularGlyph`).

This is not yet the most optimal. Ideally only the changed highlight nodes are
re-rendered.

This does mean that the highlight is undesirably considered even if RenderDag
is used outside the main graph. But most of the time the highlight is empty,
so it's probably okay.

Reviewed By: muirdm

Differential Revision: D52874611

fbshipit-source-id: 606e658d1dcdae03f097114620794ff1b94dab9f