eslint: add a rule to detect RecoilValue(map).get(key) pattern
Summary:
This pattern:

  function MaybeBackground(props: {id: string}) {
    const set = useRecoilValue(shouldUseBackgroundAtom);
    const hasBackground = set.has(props.id)
    return hasBackground ? <Background /> : null;
  }

triggers re-render of all `MaybeBackground` once the atom has any small
changes. By using a `selectorFamily` like:

  const shouldUseBackgroundById = selectorFamily({
    key: 'shouldUseBackgroundById',
    get: (id) => ({get}) => get(shouldUseBackgroundAtom).has(id),
  })

  function MaybeBackground(props: {id: string}) {
    const hasBackground = useRecoilValue(shouldUseBackgroundById(props.id);
    return hasBackground ? <Background /> : null;
  }

It can avoid re-render for unchanged items.

I have manually found and fixed several places. It seems to be a common pitfall
that is worth automatic detection. This diff adds an ESLint rule to detect the
pattern. It does so by looking for `useRecoilValue` variables, their references
in the scope, and check if all references (usage) are just `.get`, or `.has`.
`.get` inside an array function is allowed since it could be legit loops.

I first wrote the rule in a separate small project using the latest (8.56)
ESLint for easy debugging. The rule is then backported to ISL, which uses an
older ESLint and seems quite tricky to upgrade. `??` in the rule are for the
old (8.20) ESLint compatibility.

Reviewed By: evangrayk

Differential Revision: D52960285

fbshipit-source-id: 5b8ac717df0d9e054d0e463264dd7ac1b599364e