DiffBadge: further hide diff info on obsoleted commits
Summary:
The obsoleted commits should really be ignored and not operated on (including
"Ship It", "Sync", etc). This diff hides the "colorful" diff badge to make the
obsoleted commits less appealing to operate on, encouraging users to use the
successor commits instead.

The info like "Accepted" is also technically misleading. For example, the
"Accepted" state is on the latest code review version. The obsoleted commits
might be previous versions that might be not "Accepted".

The check is done at `DiffInfo`, so we can skip calculating "sync status" for
obsoleted commits too.

Reviewed By: evangrayk

Differential Revision: D52964654

fbshipit-source-id: c6d6c2f38780306e2a3b9fc10d6e86338c62e2f5