Fanouttable: Refactor to avoid 12 uninit_vec lints
Summary:
Fixes 12 instances of [`clippy::uninit_vec`](https://rust-lang.github.io/rust-clippy/master/index.html#uninit_vec).

This design of the `Fanouttable` was easy to misuse for no particular benefit since no client would ever re-use a pre-existing buffer.
Construct the `Vec` of locations inside `write` so that we don't need to pass around uninitialized memory and make every caller `unsafe`.

The new design is not ideal either, with an `Option` in and an `Option` out that are tied to each-other, but it has the advantage of being equivalent to the previous code, including in terms of performance implications, which gives confidence that it is not introducing a subtle regression.

The fix is duplicated since the code is duplicated.

Reviewed By: liubov-dmitrieva

Differential Revision: D52957379

fbshipit-source-id: 8c964d507e6edeea12aad71dc9e64daa4ec196ac