Optionally limit the number of PRs created in each submission
Currently, when submitting a PR in Sapling, it's easy to accidentally create multiple PRs when only one was intended (e.g. forgetting to run `sl pr follow`, or editing a PR that is based off a different branch).

With this change, you can configure Sapling to limit the number of PRs that can be created in a single submit action, e.g. to just 1:
```
sl config --user github.max-prs-to-create=1
```
Now when attempting to create multiple PRs in a single `sl pr submit` action, Sapling will fail with:
```
abort: refused to create 2 pull requests, max is 1
```