migrate away from tunables
Summary:
There were 3 tunables at play in this crate controlling the commit squashing:
 * megarepo_squashing_config_override_targets - the list of branches that the next tunables should be applied to
 * megarepo_override_squashing_limit - the maximum branch length for suqasting - if not overriden can be also set by client
 * megaarepo override author check - whether we check if all commits on branch are from the same author

Currently we have the list of few branches in the first, limit is set to 32 and
author check is disabled.

I'm replacing this system with two justknobs switched by "repo:branch" string.
This will allow me to emulate the current behaviour and migrate to saner
settings in the future.

Reviewed By: liubov-dmitrieva

Differential Revision: D53001224

fbshipit-source-id: f7f245047c0f4b2d69dce499e23a39db4c5cd88a