jotai: create Recoil atom in sync with Jotai atom
Summary:
To incrementally migrate from Recoil to Jotai, we need a way to "sync" between
jotai and recoil. This diff adds such a utility so we can get atoms in both
worlds in sync.

Initially I made the API return both atoms. Later I found it'd be easier to just
take a Jotai atom and return a Recoil atom. This way, if we get rid of the
Recoil atom, we can just delete code without affecting the Jotai atom
definition.

Reviewed By: evangrayk

Differential Revision: D52965494

fbshipit-source-id: 5867622610d40656f63a4825d360d8c9236e106e