jotai: migrate rest of persistAtomToLocalStorageEffect atoms to Jotai
Summary:
The interesting one is the effects on the drawerState. I migrated them to
global events instead.

I cannot get "loads state on startup" test working so I made it skip for now
(the "avatars" cache also uses the same API and the test is capturing the
avatars cache calls).

Reviewed By: evangrayk

Differential Revision: D52966669

fbshipit-source-id: 637128ff033df7f19241acfd93efbea6ef7edcb7