mononoke: Fix unused imports
Summary:
As of Rust 1.75, the compiler is better at detecting unused imports.

In particular, if something is public but not actually re-exported through the crate's interface (for instance, it's public in a module but that module is private) and not used within the crate, the compiler actually detects it as unused.

Another example is to `pub use my_mod::*;` where there is nothing public inside `my_mod` to re-export.

This diff brings us back to 0 clippy lints or warnings in `//eden/mononoke/...`

Reviewed By: liubov-dmitrieva

Differential Revision: D53000794

fbshipit-source-id: 65be30a4433325beed74a5dba38e946b99a41b4f