mononoke: Allow clippy::map_identity where used intentionally
Summary:
As of Rust `1.75`, clippy seems to be reporting `map_identity` lints more agressively.

In all of these cases, the use of this pattern is intentional as it's the simplest way to go from a ref of tuple to a tuple of refs.

Add a comment everywhere explaining why we're deciding to allow this warning.

Reviewed By: liubov-dmitrieva

Differential Revision: D52998679

fbshipit-source-id: d74d4db244fee1f5d00de3ad4aeee19f37747589