queue: support consistent work distribution across tasks
Summary:
We discovered that https://fburl.com/scuba/mononoke_derivation_worker/kwps2dc2
was interesting because we seemingly have 20 derivation tasks in prn (Oregon),
16 in nao (Ohio) and 16 in frc(Carolina) based on
https://fburl.com/code/0d5gi9ut. But the derivation log distribution suggests
60% derivations in Carolina, 36% derivations in Ohio and 4% derivations in
Oregon. We expected the logs to largely mimic the task distribution over a long
enough period if the tasks are randomly picking up the next item from the
queue.

mzr suggested this may be due to the some tasks seeing a more recent queue
state and taking upon the work before any other tasks had the opportunity to
take on the work. This diff tries to address that issue by syncing all the
clients before attempting a dequeue.

Differential Revision: D53041980

fbshipit-source-id: 4bdd58296732f29c2b8e47c0b8191ba2568c7474