Remove usages of Option<NonRootMPath>
Summary: This diff removes usages of `Option<NonRootMPath>` from walker code and replaces it with `MPath` wherever appropriate. It also comments and highlights the areas where `Option<NonRootMPath>` will continue to exist since its not used to represent paths that can be root but instead express some conditional logic based on the `Some` and `None` variants.

Reviewed By: liubov-dmitrieva

Differential Revision: D52999394

fbshipit-source-id: c016d25bef263eb31eee55cd3644b0f6edece209