Use newadmin instead of admin for bookmarks
Summary:
The scripts in mononoke_bootstrap use some env variables to get the binaries in tests and this requires replacing dots and slashes. See details [here](https://fburl.com/code/54mcnsrh).

There's a bug in the sed expression which instead of replacing literal dots for `DOT` it replaces any character.

This lead to the test breaking when I tried to use newadmin instead of admin, because admin was a special case on an if-else (`path == "."`) and newadmin wasn't.

This diff just fixes the bug in the sed expression and uses newadmin for the `bookmarks list` command.

Differential Revision: D53041086

fbshipit-source-id: 654b75ac5d2924d787ae5311bf697d27727e334a