Document usages of Option<NonRootMPath> that cannot be removed
Summary: Initially I tried to remove occurrences of `Option<NonRootMPath>` in the megarepo codebase assuming that it was used as a replacement for `MPath`. However, the integration tests indicate that `Option<NonRootMPath>` is used as a conditional flag along with the path and the code logic goes one way or the other depending on `Some` or `None` variant. Since these occurences cannot be replaced, I am documenting them so that we remember to keep them as-is.

Differential Revision: D53008665

fbshipit-source-id: 56e43632cb3d63b5c2995f9970ef4047b1ff5164