client: support not specifying region for connection
Summary:
Part of the tunables deprecation highlighted at
https://docs.google.com/document/d/1F-ekJTLZE-KK_WXCNYYHUZylDkjY_tVsfY6ifbd-cU0/edit.

We want to get rid of the "preferred_write_region" tunable which seems to be
not needed post the evolution of `derived_data_service` to support remote
derivation. In particular, even though services, such as Mononoke, make a
request to the derived data service in PRN, the derived data service itself
just queues this work to the derivation queue which the derivation workers
distributed across regions end up consuming. Therefore, the
"preferred_write_region" seems to be relic of the past where the derivation
happened within the derived_data_service itself.

Reviewed By: YousefSalama

Differential Revision: D53045089

fbshipit-source-id: 6a2ee553bcb156392fb933f3b310f8611a1e5221