Switch from CRA (webpack) to Vite (rollup)
Summary:
Change `yarn start`/`yarn build` in `isl/` to use vite (which uses rollup) instead of Create React App (which uses webpack) to build/bundle.

For clarity, webpack and rollup are bundlers.
Vite uses rollup, but also provides a dev server and hot module reloading and default configs for things like typescript and styles.
CRA uses webpack, but also provides a dev server and hot module reloading and default configs. CRA provides even more configs than Vite, including tests and more.

Vite/rollup is much faster than CRA/webpack, and requires slightly less configuration. CRA can be tough to work with, and is generally slow. Webpack is probably slightly more powerful than rollup, but migration was not all that hard so I don't think we're missing anything we need.

I did future proof slightly by  checking that vite can support StyleX (which requires a build step)

 ---

There are several aspects that need to be handled differently by migrating away from CRA.

## jest

CRA uses jest for unit tests, but it provided the config. We now need to give our own config. After copying this from the CRA template and some tweaks, all the tests pass

## entry points

We use different entry points to support different platforms. In CRA, we hacked the config to use multiple combined entry points, so that the platform definition would be inserted above the main js code. So we define e.g. androidStudio's entry point to be the concatenation of the androidStudio platform and then the main app entry point.

This notion of combined entry points does not seem to exist in rollup. Instead, we simply append a dynamic import for index.tsx at the bottom of the platform.ts.

This technically means the browser loads the index.tsx separately after loading and parsing the platform.ts. In my testing, this is a difference of 7ms compared to the other model, which I think is acceptable.

## html files

vite is a bit particular about entry point html files, and I couldn't seem to get it to work in a sub folder. So all the platform .html files are at the root, and they are mostly duplicated to each other (unlike before, where we dynamically generated them from a template). Perhaps we can clean this up in the future.

----

## Future

This only migrates the ISL client away from webpack. We still have 3 uses of webpack:
- isl-server
- vscode extension host
- vscode webview

We can separately migrate these to vite/rollup.

(Note: this ignores all push blocking failures!)

Reviewed By: quark-zju

Differential Revision: D52895088

fbshipit-source-id: 3b3a096664afa86710b7f72bff6df2540391e7b1