Change isl-server to be bundled with rollup
Summary:
More context in D52895088.

Instead of bundling the isl-server code with webpack, bundle it with rollup.

In D52895088 we use vite, which uses rollup + esbuild under the hood. Vite is intended to be a frontend bundling tool for web targets. It provides hot module reloading and a dev server. We don't need that for isl-server/. We just need a bundler, which is what the underlying rollup bundler is for. So instead of vite, we use rollup directly.

## plugins
We need several plugins to match our previous webpack behavior.

`alias`: somehow webpack was able to find isl/ and shared/ imports, probably due to it knowing about the yarn workspace. Rollup requires some helping to resolve these imports

`esbuild`: this uses esbuild for the build step, which is much faster than rollup's builtin system. Note that we could have used esbuild with webpack too, but I prefer being consistent with the front end and using rollup everywhere. The rollup team is rewriting rollup in rust ("rolldown"), coming in the near future. So rollup seems like a good place to standardize.

`nodeResolve`: we need to explicitly opt in to resolving node_modules with the node resolution algorithm. This allows it to find deps in node_modules.

`cjs`: we need to explicitly opt in to commonjs importing as well. This fixes default imports.

`terser`: this slightly optimizes bundle sizes in production. Not a huge deal optimization though. Note that we have source maps for better error messages. Frankly, we could probably just remove this for simplicity.

(Note: this ignores all push blocking failures!)

Reviewed By: quark-zju

Differential Revision: D52921975

fbshipit-source-id: e94dd6d60e233431abff9fdab566cb66605719d6