Use rollup / vite for vscode extension build
Summary:
Now that we've switched to use vite/rollup for bundling isl and isl-server, we should do the same for vscode.

This is a little annoying because of how vscode webviews work.

Firstly, we need two bundles: extension and webview. We previously handled this with two webpack configs and two webpack commands. We'll keep that pattern here.

The extension side uses rollup relatively straightforwardly like how the isl-server does it. We merely need to define vscode as an external. Otherwise, the config is similar to isl-server.

The webview side is only slightly more complicated.

We want to use vite for the webview build, to match the isl client build. However, we don't have a traditional web environment because it runs in a webview.

VS Code webviews must be loaded by defining their html in the extension. This makes it difficult to use the vite server. I actually gave this a go, by having the extension make a curl request to the vite dev server localhost:3000, then set that as the webview panel's HTML. Some further injection of values was necessary. In the end it was unfortunately very difficult to make it work, and there were several major problems.

Then I tried using rollup directly instead of vite. This was also bad because I had to redo a lot of work vite does for us, like defining how to import css files and image files, and how to copy image resources, and more.

Instead, we go with the more direct translation of using `vite --build`. This removes hot module reloading (which already wouldn't really work in webviews due to the vscode API aquisition). But by also giving `--watch` to build, we can have dev builds that hot reload.

This spits out the artifacts to the same `dist/webview` folder as before.

The big drawback here is that this is actually slower than webpack's incremental build... since what we're doing is not incremental at all. It's a full rebuild after each change. However, each full rebuild takes ~2 seconds, which is still palatable. For comparison, webpack's hot reload was ~300ms. However, webviews require a window reload ANYWAY, so hot reload and incremental rebuild time are a bit less import for vscode. We usually do development in the browser anyway.

I think it's valuable to have all of our builds use the same bundler system for consistency and simplicity of understanding.

**Is this reduction in incremental build time worth it?**

~~TODO: syntax highlighting is not yet loading, need to fix something about loading the onigurama WASM in vite.~~

(Note: this ignores all push blocking failures!)

Reviewed By: quark-zju

Differential Revision: D52981208

fbshipit-source-id: 3bd8e098a2b82546564fc5a9ce97f85172160833