Delete blobstore-fetch and blobstore-command command from old admin
Summary:
## This stack
I'll be working on migrating admin commands to newadmin.

The first few diffs will be removing commands that seem to have already been implemented in newadmin, but weren't deleted.

## This diff
Deleting the `blobstore-fetch` command. D33791723 added both fetch and upload blobstore commands, along with integration tests.

One thing I noticed is missing is the usage of `--blobstore-scrub-action` argument in the integration test. I tried adding `blobstore fetch` commands with those arguments (based on the old admin test), but the output was exactly the same, so I just reverted it because it wasn't actually testing any different behaviour.

I was planning on spending time to extend the test case to properly test the `--blobstore-scrub-action` argument, but I did a quick codesearch and it looks like this argument is already tested with both values in the `test-new-walker-scrub-blobstore.t` test ([code](https://fburl.com/code/ss1p3nps)).

markbt did I miss anything or is it safe to delete this? Should I spend time extending the tests for `--blobstore-scrub-action` or are the walker tests enough?

Reviewed By: mitrandir77

Differential Revision: D52998194

fbshipit-source-id: ac0a3994217ee68c7a4b1a3ffee7646f4b8f137d