Replace MononokePath with MPath
Summary:
The type `MononokePath` was created to represent paths that can be empty (i.e. root) and non-empty. With the introduction of `MPath`, this purpose is already served. We can now replace the occurrences of `MononokePath` with `MPath` with a few tweaks.

This diff is necessarily big since the switch needs to be atomic so we can delete `MononokePath` completely.

Reviewed By: mitrandir77

Differential Revision: D53050956

fbshipit-source-id: 962862c6452631c65164528f7595e5ff1764470a