jest: add a transformer that kills `import.meta.hot`
Summary:
Without this, `jest` stops working with upcoming changes:

 isl/src/serverAPIState.ts:40
    }), import.meta.hot);
               ^^^^
    SyntaxError: Cannot use 'import.meta' outside a module

This `import` is a keyword so it cannot be fooled by setting a `import` global variable.
I also tried configuring jest to use ESM (https://kulshekhar.github.io/ts-jest/docs/guides/esm-support)
but it does not seem to take effect. From the doc that might mean all
jest-defined globals are gone and tests need change too.

Besides, moving `import.meta.hot` to a utility module is not a choice, since
that affects the utility module, not the callsite. I believe the existing code
`declineHotReload.ts` disables hot-reloading for `declineHotReload.ts`, not for
the callsite.

So I ended up with this transformer hack. Ideally Jest just supports ESM
somehow but for now this is what unblocks next changes.

Reviewed By: evangrayk

Differential Revision: D53074125

fbshipit-source-id: 486eb1932f2836ae5afaf48dabf3f13d9e69ed89