fanouttable: Refactor away unsound loop
Summary:
The previous logic triggered a clippy warning: `while_immutable_condition`.

That was because we were looping on `start` and `end`, but never updating `start`, which meant that if we never found what we were looking for, we would be stuck in an infinite loop.

Change the logic to do what the code probably indended to do to begin with.

Apply the fix twice to duplicated code :)

Reviewed By: zzl0

Differential Revision: D53042768

fbshipit-source-id: 849384face0961a9713dedaa1c1222726f2de2c5