Test cases for recursive tag import
Summary: The test cases that we have for `gitimport` and `remote-gitimport` do not test the behaviour when dealing with recursive tags (i.e. tags pointing to other tags). I recently discovered a bug when dealing with recursive tags in whatsapp/server and had to look at the code of `gitimport` and `remote-gitimport` to understand if it works with recursive tags or not. Having these test cases will make it more explicit.

Reviewed By: gustavoavena

Differential Revision: D53098145

fbshipit-source-id: 5e47fda49b74540805c468de35e3271bbd1ec124