Update test case output to be independent of Git binary version
Summary:
We use the `git bundle verify` command in a lot of our Git integration tests. Until recently, the command would end with the text `The bundle records a complete history.`. However, the git binary on my devserver now prints another line after that `The bundle uses this hash algorithm: sha1`.

In general, its not a good idea to depend on the verion of the binary for the format of the output. Hence, I am updating the integration tests to use the `-q` flag which will always only output information about the validity of the bundle (the rest of the information was pretty much all glob anyway)

Reviewed By: clara-9

Differential Revision: D53124589

fbshipit-source-id: 0e87e6f348353eba9e1eb1504d99e09001645eae