Seed documentation for Live Git Sync
Summary:
Set-up a docs folder under `eden/mononoke/git`, configure it so that the wiki will by synced to the docs folder in the code.

Seed the page with one paragraph.
I want to confirm that the configuration works properly by landing this diff before writing the rest of the documentation.

Reviewed By: YousefSalama

Differential Revision: D53125597

fbshipit-source-id: b07b9e93934432f1480e70b53bf49907534fad0d