Test case highlighting gitbundle failure in face of recreated recursive tag
Summary:
The current set of test cases for `gitbundle` do not include scenarios with recursive tags. I have added two variants of recursive tags through this new test case. The first variant is regular recursive tags, i.e. tags that point to other existing tags. This is not recommended or standard behavior in Git, but nevertheless it is possible to do so. Our current implementation supports this variant. The other (crazy) variant is rewriting an existing tag to point to the previous version of itself instead of some commit. For example, if we had a tag `some_tag` that pointed to commit `some_commit`, we can reset it to point to the git object which represents the existing version of `some_tag`.

This is an abnormal scenario and doesn't provide any meaningful functionality, but for some reason such a pattern exists in `whatsapp/server` which is why we need to cater to it. Right now, having such a tag leads to incorrect bundle creation which in turn leads to failed git clone. Once I fix this, I will update the test accordingly.

Reviewed By: gustavoavena

Differential Revision: D53125609

fbshipit-source-id: 3d7b468508349e2f7dfe89e339b8eaefbf6e996c