Extend bonsai_tag_mapping schema to support target_is_tag flag
Summary:
A highlighted in the previous diff in the stack, there is a case where we miss out on including recursive tags in the bundle or when pulling/cloning through Git protocol. This diff adds a new `target_is_tag` flag to the `bonsai_tag_mapping` table which will indicate if the tag is a recursive tag (i.e. tag pointing to a tag). We can then use this flag at pack creation time to figure out if we need to parse the tag object and include its target. If we did not have this flag, we would have to parse every single tag object during pack creation which would be a very expensive operation.

Note: Will land this only after D53127921 has landed

Reviewed By: markbt

Differential Revision: D53127120

fbshipit-source-id: 5c03676efe3991251510e6d2e82fcef72e384ca7