Delete cxx_bridge @rust workaround
Summary:
From D24313096 — this was a relatively disruptive workaround for a Buck1 bug that we never fixed.

https://www.internalfb.com/code/fbsource/[cf6644a2c08ba692a7b16ca53d62136b2ef503bb]/tools/build_defs/fbcode_macros/build_defs/rust_library.bzl?lines=83-93

In Buck2, Python depending on Rust is correctly handled, regardless of whether there is any C++ target in between.

There are still some usages of `cxx_bridge` in fbsource/arvr, which possibly build with Buck1. Hopefully they do not involve libomnibus.so, in which case they do not require the workaround. Libomnibus is deprecated in favor of Native Python anyway. If needed, we can add back the rust workaround with an `if not is_buck2()`.

Reviewed By: zertosh, shayne-fletcher, JakobDegen

Differential Revision: D53095236

fbshipit-source-id: def15aac783a921b87b66a3c2d1ccdc5ec20f8f8