Don't include vscode webview source maps in production
Summary:
vscode webviews don't load external (.js.map) source maps automatically. So there's no real point in including source maps for the webview.

I added some instructions to the readme for how to manually use a source map.

We can produce source maps for dev builds so we can do this manual process. But in production, we can just use a dev build to get the source map, instead of having the user have a copy that they can't use.

Reviewed By: quark-zju

Differential Revision: D53146220

fbshipit-source-id: 92ba33afa8141d9dc1e2e9136df2894cb737f786