git: Document Live git sync end-to-end
Summary:
There are probably sections where we could elaborate a bit more, but we are now documenting the entire process of setting up live sync.

Note: Compared to how we started, we now always recommend to set-up new live synced repos as SCS tier and not prod tier.
This means we don't treat repos with or without submoduled differently and we keep a narrower contract with our users.

Reviewed By: singhsrb

Differential Revision: D53186311

fbshipit-source-id: f54d05cb35fa2cf29dfc5e664af1a57c610e6dc3