Revert D53095236: Multisect successfully blamed "D53095236: [fbcode_macros] Delete cxx_bridge @rust workaround" for otest failure
Summary:
This diff is reverting D53095236
D53095236: [fbcode_macros] Delete cxx_bridge rust workaround by dtolnay has been identified to be causing the following test failure:

Tests affected:
- [fblearner/flow/projects/modelstore/tests:pyper_publish_integration_test - test_pyper_offline_publish_gpu_and_acc (fblearner.flow.projects.modelstore.tests.pyper_publish_integration_test.PyperPublishIntegrationTest)](https://www.internalfb.com/intern/test/562950022580873/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4169370
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: dtolnay

Differential Revision: D53189119

fbshipit-source-id: 7942a30c57e858d5e8cbe5c8b0706f550224b9b9