Fix entangled atoms not being initialized if written too soon
Summary:
We saw an issue where the "pull" button would disappear when closing & reopening the ISL tab in vscode. This happened because repositoryInfo was undefined, which was happening because the atom was getting written before the <RecoilRoot> renders and sets up globalRecoil.

I think the close&reopen repro happens because some messages send over the vscode API happen more quickly the second time ISL is opened (didn't fully dig into this)

Reviewed By: quark-zju

Differential Revision: D53193805

fbshipit-source-id: a2f2ab5ccf54678bc1562363e1b4c80152fefdf0