journal: oxidize journal entry
Summary:
We'll make use of this when implementing basic functionality of the journal in Rust. This moves the struct plus the serialization/deserialization of journal entries into Rust.

This also makes journal entries always have integer timestamps, as unix timestamps with decimals seems to be an unintended consequence of our Python 2 -> 3 migration

Reviewed By: muirdm

Differential Revision: D52954922

fbshipit-source-id: a62dc5e6b5ed4777445eb345bcab3c31d6d305ad