Rust CLI: fix getSnapshot for FilteredFS repos
Summary:
# Problem:
FilteredFS repos write FilteredRootIds to the Snapshot file so that they can track the current working copy parent, active FilterID, and parse those two pieces of information as needed.

This means that the current getSnapshot logic naively returns FilteredRootIds instead of the expected commit hashes as expected.

# Solution:
We can modify getSnapshot to properly parse the FilteredRootIds and return only the commit hash to the caller.

Reviewed By: jdelliot

Differential Revision: D53146665

fbshipit-source-id: c38980f00b977f525fba654e146e88f8e1693c43