Rust CLI: refactor config deserialization
Summary:
# Problem:
The next diff implements correct getSnapshot() logic which requires doing a comparison of the repo_type config field. Instead of comparing raw strings, we should compare enums.

# Solution:
We can parse the repo_type CheckoutConfig option as an enum to avoid string comparison for determining repo types.

This requires implementing to/from string implementations for the new RepositoryType enum. We can use the strum crate to do this for us.

We can then use the to/from string implementations to deserialize (and serialize) CheckoutConfigs.

Reviewed By: kmancini

Differential Revision: D53146609

fbshipit-source-id: 29cc702695381021b210f85f83855b84dc11534e