Implement deleted-manifests command in newadmin
Summary:
## This stack
See summary in D53181324.

## This diff
Implements the `deleted-manifests tree` subcommand.
There are 2 other subcommands: `fetch` and `verify`. I believe that the first one is already supported by the `blobstore fetch` command with the `--decode-as` flag, so I won't port it.

The `verify` command  for deleted manifests still need to be implemented in the `verify-manifest` newadmin command.

Reviewed By: liubov-dmitrieva

Differential Revision: D53183145

fbshipit-source-id: 0da0873c263a433c8d5375b56b6b97ec6d122581