config: move dynamicconfig refresh later in config loading
Summary:
Separate the dynamic config background refresh and move it later. This allows it to run with the full config at its disposal (for example to fetch the "configs.generationtime" value that controls the mtime check).

Previously the config refreshing ran in an intermediate state that basically only had the existing dynamic config loaded. This is incorrect because it was missing most of the config sources (e.g. static, user, repo, etc.), and has the wrong priority since the "--config" CLI options were being overwritten by dynamic config values.

Reviewed By: zzl0

Differential Revision: D53195183

fbshipit-source-id: 91717b6be67a885304b4595f31e4832103bf3642