report APFS purgeable space as user-fixable
Summary:
## Problem

Having an "automatic remediation" for full storage is misleading:

1) Our remediation may not actually solve the full disk space issue
2) Our remediation doesn't actually work since the `apfs.util` command we run requires `sudo` to correctly purge all purgeable space.

## Solution

Instead of claiming that we automatically remediated the issue, let's inform the user of the potential problem, tell them how they can check if that's the *actual* problem, and then give them a remediation they can run to fix the problem. While this is strictly worse than automatically running the remediation for them, it's much better than running a remediation that doesn't actually work and claiming it fixed the problem.

Reviewed By: genevievehelsel

Differential Revision: D53200901

fbshipit-source-id: 23dc142b4c0d236d8b31bba85d0a8f98fe76aeba