Allow additional metadata to be associated with fetched refs
Summary: During the negotiation phase of Git protocol, we not only need to send the list of refs/symrefs and the commits that they map to but also apped some additional metadata to those based on the output. This diff updates `generator.rs` in the protocol crate to support these additional demands when generating packfile and bundle. A follow up diff will add a new generator specifically for the PackHeader data using these new parameters.

Reviewed By: liubov-dmitrieva

Differential Revision: D53189062

fbshipit-source-id: 1aab257a101f2ffc7d43c8ca0e4e827bdc7c4eb4