contrib: add a script to print Rust deps
Summary:
We have lots of small dependencies. Manually write the Rust dependencies can be annoying.
This 100-line script uses some naïve regex to find out the dependencies and
print them as a list that can be copy-pasted to `TARGETS`. It does not do any
AST parsing and is not aware with more complex features like`#[cfg(test)]` so
it is inaccurate. But it provides some "initial" list that seems still helpful.

Reviewed By: muirdm

Differential Revision: D53211953

fbshipit-source-id: 03714c730d7e6c00ce4a77dbcd2985893ce73d56