commands: extract HgPython to a separate cmdpy crate
Summary:
Some commands implementation dependents on `HgPython`. To prevent those
commands from depending on the `commands` crate and cause a circle, let's move
the `HgPython` part out to a separate crate.

It seems this somehow changes the modules used in tests to be from static not
filesystem. I didn't investigate deeper about why that happened... tests are
updated to the new outputs.

Reviewed By: muirdm

Differential Revision: D53205905

fbshipit-source-id: c65777ea9ce9e8bbd6cf5f9d17618a129d957a26