Add StrictUnion type helper
Summary:
This is a small type utility that makes it a little easier to work with data from graphql.

We get types back from graphql like:
```
type TypeaheadResult = {
  __typename: 'diff',
  diff: {
    title: string
  }
} | {
  __typename: 'task',
  task: {
    title: string,
  }
}
```

This means when looking at the results, e.g. filtering, you need to cast:
```
results.map(result => {
  return (result as typeof result & {__typename: 'diff'}).diff?.title ?? (result as typeof result & {__typename: 'task'}).task?.title;
});
```
otherwise, since __typename is not narrowed at all, both the task and diff fields are missing, as opposed to optional.

This type util converts the union into one where all the keys are required (but nullable).
So this type becomes:
```
type TypeaheadResult = {
  __typename: 'diff' | 'task',
  diff?: {
    title: string
  },
  task?: {
    title: string
  }
}
```

You still have to validate that you have the property you want, but this is easy with null coalescing:
```
results.map((result: StrictUnion<Result>) => {
  return result.diff?.title ?? result.task?.title;
});
```

Reviewed By: quark-zju

Differential Revision: D53277531

fbshipit-source-id: 005cb7681c79bf197689fd8db1a74a9cc5684e7d