Trim commit description from SSL (#820)
Summary:
Trim commit description from SSL

Currently, when a commit is created with a title/description, the PR/commit description is:
```
{title}

{desc}
```

When the description was parsed out previously, the first line (and trailing linebreak) were stripped, but the line in between title/desc was left around. This meant every time you amended a commit/PR, another linebreak got added to the top.

I figure its probably fine to just trim whitespace from the description in general, but could update this to only trim leading whitespace (cant really just change the slice() index to ` + 2` because the commit may come from another source than sapling, and may not contain this extra linebreak)

Pull Request resolved: https://github.com/facebook/sapling/pull/820

Reviewed By: quark-zju

Differential Revision: D52937974

fbshipit-source-id: 0b1fbb77d29fd0c17bf5631a15cd496245af4e3b