Remove node events dependency from TypedEventEmitter
Summary:
TypedEventEmitter ended up being useful on the client and server. But it was using `events`, which meant the client had to polyfill node behavior. That's pretty weird, and although it seemed to happen automatically in vite, I think it's better to not need server polyfills on the client if possible.

These days, browser and node have EventTarget, which is basically EventEmitter. If we're a little crafty, we can make TypedEventEmitter work.

This was also causing some weird issues in buck builds...?

Reviewed By: quark-zju

Differential Revision: D53293331

fbshipit-source-id: a8e1590732dff9300f3f12a247a2d4be11bcc535