Show alerts
Summary:
Add a way of showing alerts for ongoing issues in the product. This is controlled by a section of config settings, like so:

```
[alerts]
S11111.title=My Alert
S11111.description=What's wrong
S11111.url=https://sapling-scm.com
S11111.severity=SEV 1
S11111.show-in-isl=true
```

This config is expected to be set dynamically to notify users.

We show alerts from this config at the top of the UI. If you dismiss an alert, we remember that you've dismissed it and don't show that alert again.

Reviewed By: quark-zju

Differential Revision: D53249140

fbshipit-source-id: 32c71597a2b523a8fba932a7bdd96d243c6a51f7